refactor: CarRace setter와 Inputview 역할 리팩토링
